Closes #90 - filter for scala.js
Implementing a featured/supported tag cloud which contains all found
Scala versions (major.minor) and also scala-js versions. On Top of that
there are featured artefacts like spark and akka. This tag will list
all artefacts which depends on aka or spark.

Also fixes a bug with tag cloud if there is a “-“ in the tag. Now they
are displayed correct.