Makes the organization page redirect to search
This reduces duplication on the site by making it so there are no longer two places to search for an organization's projects.

This makes it less likely that links to "/organization" will proliferate, leaving that path available (less likely to break external references) if an organization page is ever added.