Add twitter summary cards for project pages
Introduces a "HeadMeta" model to add additional meta tags to the main template.
The project template uses it to inject the meta tags for a twitter card.

Only adds a twitter card if the project has an associated GitHub project and if
this GitHub project has a short description.  Uses the avatar of the GitHub
user/org as image if any.

Closes #239.