Fix search relevance
Please see https://github.com/scalacenter/scaladex/issues/389#issuecomment-490584056

From looking at GitHub's search result, which is very good, the signal we should be boosting more is the description. The tag seems to mix in more random things, so it's ok as a tie breaker, but it's not very reliable.