use scala-arm to manage WSClients
scala-arm makes resource management safer and easier.
The number of WSClient instances should stay low, and
so it is likely that you want to create one and then use it at different
places in your code. Using scala-arm minimize the risk that the resource
is closed too soon or never.