Move DataRepository to search module
The search module contains the implementation related
to elasticsearch indexing and searching.
In particular, I want to create a mapping (an elasticsearch schema)
that is different from the API data model. Because
I want to separate the dependency and the release mapping
The DataRepository will make the conversion
from the search mapping to the API model and vice versa.