Additional badge testcase where not all platform editions are supported
Prompted by https://github.com/scalacenter/scaladex/pull/660#discussion_r629233719,
this is a testcase for when Scala platform editions are not available for
all cited Scala versions. In this situation, there are a few options for
what the badge text could be:

* Over promise: `7.1.0 (Scala 3.0.0-M3, 2.13 - Native 0.4+0.3)`, which
  incorrectly implies that there's a `3.0.0-M3` version for `Native 0.4`.
* Provide full detail:
  `7.1.0 (Scala 3.0.0-M3 - Native 0.4, Scala 2.13 - Native 0.3)`.
* Omit platform versions that don't have full support for the listed
  Scala language versions: `7.1.0 (Scala 3.0.0-M3, 2.13)`.

The endpoint `latest-by-scala-version.svg`, is intended to list all
supported Scala versions, with the latest artifact version against them,
and giving that information concisely takes priority over providing
completely comprehensive information about the differing versions of the
Scala platforms. It's also best not to over-promise, so of the three
options I think the last is the best.