Move badge display to it's own box on RHS bar
Suggested in https://github.com/scalacenter/scaladex/pull/670#issuecomment-851336668

Advantages
----------

* If the project already has the badge in it's project README, then it
  would be repetitious to show the badge immediately above it in the
  project header.
* If we consider the visible badge on the project page as just a link to
  the markdown-generator lightbox, rather than informational in itself
  (given that version information is already visible in the versions
  dropdown), then there is no need to prominently display the full badge.

Disadvantages
-------------

* Narrower space usually truncates the (very wide) badge. Mitigating
  against this is the argument that the badge here is just a link, rather
  than informational.
* Takes up more space vertically, so the consumed screen-area is
  somewhat larger than with the badge in the project header, even though
  it's narrower!


It would be nice if there was a visual indication that the badge was
truncated on the right. Using `overflow: scroll` gives an always-present
scrollbar, which I find visually cluttered and more than what's needed- we
don't really want to scroll the badge from side-to-side, when we can just
click on it to view the whole thing.