Better support for sbt artifacts in the Scaladex badge
Each release of sbt uses a specific version of Scala (1.0 uses Scala 2.12,
0.13 uses Scala 2.10), so the sbt version of a plugin fully determines the
Scala language binary version it uses - this means, for a user, it's only
really interesting to know the sbt version of the plugin, not the Scala
language version.

At the moment, the Scaladex badge for, eg, the `sbt-assembly` plugin reads
`1.0.0 (Scala 2.12, 2.10)` - for the user, it would be more helpful if it
read `1.0.0 (sbt 1.0, 0.13)` - listing the versions of sbt you can actually
use it with! Although the Scaladex badge did have some logic to display the
platform versions supported by _all_ listed Scala language versions, that
doesn't help in this situation, as the Scala language version differs for
each platform version of sbt. See also this comment from a prior PR:

https://github.com/scalacenter/scaladex/pull/670#issuecomment-853426504

This commit updates the Scaladex badge logic to have two version-summary
strategies:

* `SummariseLanguageVersion` - which was the existing strategy, listing
  notable Scala version support, secondarily adding in a summary of
  platform editions (eg 'Scala 2.13 - Native 0.4+0.3')
* `SummarisePlatformEdition` - the new strategy, which is used if *all*
  fetched `ScalaTarget`s for an artifact have a `targetType` for which
  the platform version fully determines the Scala version (as is the case
  for sbt). It omits the Scala version from the summary, and just relates
  the platform versions (eg `sbt 1.0, 0.13`)