Adding license field to GithubInfo and GithubModel
Now extracting the `spdx_id` from the JSON response from GitHub as
`Option[String]`, assigned to field `licenseName` in `GithubModel.Repository`.
In order to extract a value of type `Option[License]` from this, there needs
to be a way for `GithubClient` to get a hold of `LicenseCleanup`.

I don't think there's a clean way to do this right now w/o injecting the
`LicenseCleanup` instance directly, or passing down the `DataPath`. I don't
want to do this because it seems like these two components shouldn't really be
tightly coupled together.